requirements.txt
# API server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# AWS
boto3>=1.34.0

# LLM
openai>=1.0.0

# Gemini image generation
google-genai>=1.0.0

# Runway video generation
runwayml

# Video processing
moviepy>=2.0.0

# Utilities
natsort
Pillow
requests
tqdm
